requirements.txt
# Core Flask packages
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5

# Authentication
Flask-Login==0.6.3

# Forms and validation
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0

# Security
python-dotenv==1.0.0

# Production database (PostgreSQL)
psycopg2-binary==2.9.9

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2

# Production server
gunicorn==21.2.0

# Utilities
click==8.1.7
